Moved application-specific code away from release-controller to application
